vMPU: Rename "freescale_k64" into "kinetis"
The vMPU drivers refer to the Kinetis MPU, not the Freescale K64F board
specifically. This is also consistent with the rest of the configuration
parameters in this repo, where we refer to the Kinetis MPU.
