+++ core/vmpu/inc/vmpu_kinetis.h
