+++ core/vmpu/inc/vmpu_kinetis_aips.h
