+++ core/vmpu/inc/vmpu_kinetis_mem.h
